# openapi.yaml
openapi: 3.0.0
paths:
...
  /api/collaborators:
    post:
      operationId: CollaboratorsController_create
      parameters: []
      responses:
        "201":
          description: ""
      tags: &a1
        - Collaborators
      security: &a2
        - bearer: []
    get:
      operationId: CollaboratorsController_search
      parameters: []
      responses:
        "200":
          description: ""
      tags: *a1
      security: *a2
  /api/collaborators/{cpf}:
    get:
      operationId: CollaboratorsController_getByCpf
      parameters: []
      responses:
        "200":
          description: ""
      tags: *a1
      security: *a2
    patch:
      operationId: CollaboratorsController_update
      parameters: []
      responses:
        "200":
          description: ""
      tags: *a1
      security: *a2
  /api/collaborators/{cpf}/status:
    patch:
      operationId: CollaboratorsController_updateStatus
      parameters: []
      responses:
        "200":
          description: ""
      tags: *a1
      security: *a2
...
info:
  title: Atestados API
  description: API de atestados médicos
  version: 1.0.0
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
...